# backend/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0

  title: D4com API

  description: A simple API to do CRUD on user, orders, product, upload resource
servers:
  - url: http://localhost:5000/
...
paths:
...
  /api/products:
    get:
      security:
        - bearerAuth: []
      summary: Lists all the products
      tags: [Products]
      responses:
        "200":
          description: The list of products.

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Products"
    post:
      security:
        - bearerAuth: []
      summary: Creates a product object.

      tags: [Products]

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Products"
      responses:
        "201":
          description: Created
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Products : 
      type : object 
      required : 
        - user
        - name 
        - image
        - brand
        - category 
        - description
        - reviews 
        - rating
        - numReviews
        - price
        - countInStock
      properties :
        user:
          type: string

          description: The email of the user.
        name:
          type: string

          description: The email of the user.
        image:
          type: string

          description: The email of the user.
        brand:
          type: string

          description: The email of the user.
        category:
          type: string

          description: The email of the user.
        description:
          type: string

          description: The email of the user.
        reviews:
          type: string

          description: The email of the user.
        rating:
          type: string

          description: The email of the user.
        numReviews:
          type: number

          description: The email of the user.
        price:
          type: number

          description: The email of the user.
        countInStock:
          type: number

          description: The email of the user.
      example : 
        rating: 3.5
        numReviews: 2
        price: 405
        countInStock: 30
        id: "61dc27d36ded5d3ce4522d5e"
        name: "OnePlus Nord 2 5G"
        image: "/images/op_nord2.jpg"
        description: "Blue Haze color, MediaTek Dimensity 1200, 8GB RAM, 128GB Storage, 4500mAh battery, Android 11, 50MP Triple Camera"
        brand: "OnePlus"
        category: "Mobiles"
        user: "61dc27d36ded5d3ce4522d4f"